Works for uninitialized dekstops

+++ showHideDesktop.sh
@@ -1,13 +1,22 @@
 #!/bin/bash
 
-STATUS=$(defaults read com.apple.finder | grep CreateDesktop | sed 's/[^0-9]//g')
+defaults read com.apple.finder | grep CreateDesktop > /dev/null
 
-if [ $STATUS == 0 ]
+if [ $? == 0 ]
 then
-	defaults write com.apple.finder CreateDesktop -bool true
-	killall Finder
+	
+	STATUS=$(defaults read com.apple.finder | grep CreateDesktop | sed 's/[^0-9]//g')
+	
+	if [ $STATUS == 0 ]
+	then
+		defaults write com.apple.finder CreateDesktop -bool true
+		killall Finder
+	else
+		defaults write com.apple.finder CreateDesktop -bool false
+		killall Finder
+	fi
 else
-	defaults write com.apple.finder CreateDesktop -bool false
+	defaults write com.apple.finder CreateDesktop -bool true
 	killall Finder
 fi
